requirements.txt
einops==0.8.1
pandas==2.2.1
pydantic==2.7.1
torch==2.5.1
torchmetrics==1.6.3
typer==0.15.2
lightning==2.5.0
pillow==11.2.1
pyarrow==20.0.0
loguru==0.7.2
scikit-learn==1.7.0
timm==1.0.15
transformers==4.51.3
wandb==0.16.5
